requirements.txt
numpy==2.2.6
pygame-ce==2.5.6
pytest==9.0.2
